Extract QueueButton as component
