+++ src/components/actionbar/ComfyActionbar.vue
@@ -8,48 +8,7 @@
     <div class="actionbar-content flex items-center" ref="panelRef">
       <span class="drag-handle cursor-move mr-2 p-0!" ref="dragHandleRef">
       </span>
-      <div class="queue-button-group flex">
-        <SplitButton
-          class="comfyui-queue-button"
-          :label="activeQueueModeMenuItem.label"
-          :icon="activeQueueModeMenuItem.icon"
-          severity="primary"
-          @click="queuePrompt"
-          :model="queueModeMenuItems"
-          data-testid="queue-button"
-          v-tooltip.bottom="$t('menu.queueWorkflow')"
-        >
-          <template #item="{ item }">
-            <Button
-              :label="item.label"
-              :icon="item.icon"
-              :severity="item.key === queueMode ? 'primary' : 'secondary'"
-              text
-              v-tooltip="item.tooltip"
-            />
-          </template>
-        </SplitButton>
-        <BatchCountEdit />
-        <ButtonGroup class="execution-actions flex flex-nowrap">
-          <Button
-            v-tooltip.bottom="$t('menu.interrupt')"
-            icon="pi pi-times"
-            :severity="executingPrompt ? 'danger' : 'secondary'"
-            :disabled="!executingPrompt"
-            text
-            @click="() => commandStore.execute('Comfy.Interrupt')"
-          >
-          </Button>
-          <Button
-            v-tooltip.bottom="$t('sideToolbar.queueTab.clearPendingTasks')"
-            icon="pi pi-stop"
-            :severity="hasPendingTasks ? 'danger' : 'secondary'"
-            :disabled="!hasPendingTasks"
-            text
-            @click="() => commandStore.execute('Comfy.ClearPendingTasks')"
-          />
-        </ButtonGroup>
-      </div>
+      <ComfyQueueButton />
       <Divider layout="vertical" class="mx-1" />
       <ButtonGroup class="flex flex-nowrap">
         <Button
@@ -68,78 +27,20 @@
 import { computed, nextTick, onMounted, ref, watch } from 'vue'
 import Panel from 'primevue/panel'
 import Divider from 'primevue/divider'
-import SplitButton from 'primevue/splitbutton'
 import Button from 'primevue/button'
 import ButtonGroup from 'primevue/buttongroup'
-import BatchCountEdit from './BatchCountEdit.vue'
-import {
-  AutoQueueMode,
-  useQueuePendingTaskCountStore,
-  useQueueSettingsStore
-} from '@/stores/queueStore'
-import { app } from '@/scripts/app'
-import { storeToRefs } from 'pinia'
+import ComfyQueueButton from './ComfyQueueButton.vue'
 import { useSettingStore } from '@/stores/settingStore'
 import { useCommandStore } from '@/stores/commandStore'
-import { MenuItem } from 'primevue/menuitem'
-import { useI18n } from 'vue-i18n'
 import { useDraggable, useEventListener, useLocalStorage } from '@vueuse/core'
 import { debounce, clamp } from 'lodash'
 
 const settingsStore = useSettingStore()
 const commandStore = useCommandStore()
-const queueCountStore = storeToRefs(useQueuePendingTaskCountStore())
-const { mode: queueMode } = storeToRefs(useQueueSettingsStore())
 
 const visible = computed(
   () => settingsStore.get('Comfy.UseNewMenu') === 'Floating'
 )
-
-const { t } = useI18n()
-const queueModeMenuItemLookup: Record<AutoQueueMode, MenuItem> = {
-  disabled: {
-    key: 'disabled',
-    label: 'Queue',
-    icon: 'pi pi-play',
-    tooltip: t('menu.disabledTooltip'),
-    command: () => {
-      queueMode.value = 'disabled'
-    }
-  },
-  instant: {
-    key: 'instant',
-    label: 'Queue (Instant)',
-    icon: 'pi pi-forward',
-    tooltip: t('menu.instantTooltip'),
-    command: () => {
-      queueMode.value = 'instant'
-    }
-  },
-  change: {
-    key: 'change',
-    label: 'Queue (Change)',
-    icon: 'pi pi-step-forward-alt',
-    tooltip: t('menu.changeTooltip'),
-    command: () => {
-      queueMode.value = 'change'
-    }
-  }
-}
-
-const activeQueueModeMenuItem = computed(
-  () => queueModeMenuItemLookup[queueMode.value]
-)
-const queueModeMenuItems = computed(() =>
-  Object.values(queueModeMenuItemLookup)
-)
-
-const executingPrompt = computed(() => !!queueCountStore.count.value)
-const hasPendingTasks = computed(() => queueCountStore.count.value > 1)
-
-const queuePrompt = (e: MouseEvent) => {
-  const commandId = e.shiftKey ? 'Comfy.QueuePromptFront' : 'Comfy.QueuePrompt'
-  commandStore.execute(commandId)
-}
 
 const panelRef = ref<HTMLElement | null>(null)
 const dragHandleRef = ref<HTMLElement | null>(null)
@@ -227,10 +128,5 @@ useEventListener(window, 'resize', adjustMenuPosition)
 
 :deep(.p-panel-header) {
   display: none;
-}
-
-.comfyui-queue-button :deep(.p-splitbutton-dropdown) {
-  border-top-right-radius: 0;
-  border-bottom-right-radius: 0;
 }
 </style>
